@@ -1,22 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Row, Col } from 'reactstrap';
-import { Link } from 'react-router-dom';
+// import { Link } from 'react-router-dom';
 
 const Error = ({ title, content }) => (
-  <Row>
-    <Col lg="4">
-      <h2>
-        {title}
-      </h2>
-      <p>
-        {content}
-      </p>
-      <p>
-        <Link to="/" className="btn btn-primary">
-          Go Home
-        </Link>
-      </p>
+  <Row className="py-5 my-5">
+    <Col md={{ size: 4, offset: 4 }}>
+      <h2 className="text-center">{title}</h2>
+      <p className="text-center">{content}</p>
+      {/* <p className="text-center"><Link to="/" className="btn btn-primary">Go Home</Link></p> */}
     </Col>
   </Row>
 );
